+++ changelog.md
@@ -1,3 +1,6 @@
+# Next
+- [BUG] Fixed issue with empty `include.where`
+
 # 2.0.0
 - [BUG] Fixed `field` support for `increment` and `decrement`.
 - [FEATURE/BUG] Raw queries always return all results (including affected rows etc). This means you should change all promise listeners on `sequelize.query` to use `.spread` instead of `.then`, unless you are passing a query type.